--- src/main/java/finalmission/member/application/in/SignupMemberIn.java
@@ -1,4 +1,0 @@
-package finalmission.member.application.in;
-
-public record SignupMemberIn(String name) {
-}
