+++ src/main/java/finalmission/member/application/in/dto/SignupMember.java
@@ -1,0 +1,4 @@
+package finalmission.member.application.in.dto;
+
+public record SignupMember(String name) {
+}
