+++ src/main/java/finalmission/popupstore/application/in/PopupStoreController.java
@@ -1,6 +1,7 @@
-package finalmission.popupstore.application;
+package finalmission.popupstore.application.in;
 
-import finalmission.popupstore.dto.CreatePopUpStoreIn;
+import finalmission.popupstore.application.PopupStoreService;
+import finalmission.popupstore.application.in.dto.CreatePopUpStore;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,7 +18,7 @@ public class PopupStoreController {
     public final PopupStoreService popupStoreService;
 
     @PostMapping
-    public ResponseEntity<Void> open(@RequestBody final CreatePopUpStoreIn request) {
+    public ResponseEntity<Void> open(@RequestBody final CreatePopUpStore request) {
         popupStoreService.open(request);
 
         return ResponseEntity
