+++ src/main/java/finalmission/shopkeeper/application/ShopkeeperService.java
@@ -1,7 +1,8 @@
 package finalmission.shopkeeper.application;
 
+import finalmission.shopkeeper.application.in.dto.CreateShopkeeper;
+import finalmission.shopkeeper.application.out.ShopkeeperRepository;
 import finalmission.shopkeeper.domain.Shopkeeper;
-import finalmission.shopkeeper.dto.CreateShopkeeperIn;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -11,7 +12,7 @@ public class ShopkeeperService {
 
     public final ShopkeeperRepository shopkeeperRepository;
 
-    public void signUp(final CreateShopkeeperIn command) {
+    public void signUp(final CreateShopkeeper command) {
         shopkeeperRepository.save(
                 Shopkeeper.create(command.name())
         );
