+++ src/main/java/finalmission/shopkeeper/application/in/dto/CreateShopkeeper.java
@@ -1,0 +1,4 @@
+package finalmission.shopkeeper.application.in.dto;
+
+public record CreateShopkeeper(String name) {
+}
