--- src/main/java/finalmission/shopkeeper/dto/CreateShopkeeperIn.java
@@ -1,4 +1,0 @@
-package finalmission.shopkeeper.dto;
-
-public record CreateShopkeeperIn(String name) {
-}
